Break the transfer note integrationo tests into their own file
